CONSOLE APP STRINGS AND INTEGERS SUB ASSIGNMENT
Minor edit that corrected an error in which my program would not interate the entire integer list.  Error caused by the fact that I had originally nested my user data entry line of code inside my foreach loop.

To correct the error, I had to move the user data entry line of code outside of the foreach loop, however, the below statement and the foreach loop are still nested in the try statement.